requirements.txt
pydantic~=2.6.3
uvicorn~=0.28.0
fastapi~=0.110.0
PyYAML~=6.0.1
openai~=1.13.3
PyGithub~=2.2.0
pytest~=8.1.1
httpx~=0.27.0
jsmin~=3.0.1
pygments~=2.17.2
tenacity~=8.2.3
